+++ jgclark.Dashboard/CHANGELOG.md
@@ -1,6 +1,19 @@
 # What's changed in 🎛 Dashboard plugin?
 For more details see the [plugin's documentation](https://github.com/NotePlan/plugins/tree/main/jgclark.Dashboard/).
 
+## [Perspectives.a22 = 2.1.0.a22] @jgc, 2024-11-???
+- (jgc): Nudge top of settings dialog down slightly, to stop it from sometimes obscuring the current Perspective name.
+- (jgc): Add new 'Moving/Scheduling Items' heading in Settings dialog
+- (dbw): Lots of fixes and improvements to dev logging for React
+- (dbw): Fix switch to clean perspective when modified
+- (dbw): Added 'Rename Perspective...' action to UI
+- (jgc): Added renamePerspective handler action to back end -- FIXME: race conditions, sigh
+- (dbw): Changed cleanDashboardSettings() so perspectives is only global setting
+- (jgc): added focus of DD input box for renamePerspective
+- (jgc): tweaked and simplified DynamicDialog header layout FIXME: but some CSS is appearing from nowhere that I can find, and so DD title is still too narrow.
+- (jgc): started changing the section addButtons to use DynamicDialog not command bar -- TODO: need advice from DBW about handleCommandButtonClick and how it propagates. Also see Dashboard.jsx's long-standing TODO comments about refreshTimer()s.
+- (jgc): fix to {s} appearing in section description
+
 ## [Perspectives.a21 = 2.1.0.a21] @jgc, 2024-11-23
 - (jgc): remove the 'JSON control' and tidy up the area in the settings dialog around where it used to be.
 - (jgc): updated README with basic info on Perspectives, ready for limited alpha testing.
@@ -19,7 +32,7 @@ For more details see the [plugin's documentation](https://github.com/NotePlan/pl
 - (dbw): fix to add perspective
 - (dbw): fix to delete perspective
 - (dbw): fixes to new console log view
-- (dbw): fixes to PerspectiveSelector
+- (dbw): Make sure dash persp is first in displayed list
 - (dbw): fix babel to work for jest and rollup for the app
 - (dbw): improve logging
 - (jgc): change priority detection to ignore ! not at start of content